refactor(models): house run_tool_loop under models/utils
Move the shared tool-use loop primitive into a models/utils sub-package and
update its logger name accordingly.
